moved form for uploading photo into separate component. #1
